fix delete with split update (#15288)
If DELETE operator is generated by SplitUpdate node (SplitUpdate
node is generated for updating the partition keys), we raise an error.

The root cause is SplitUpdate will split the origin tuple to two tuples,
one is for deleting and the other one is for inserting. we can skip
the deleting tuple if we found the tuple is updated concurrently
by another transaction, but it is difficult to skip the inserting
tuple, so it will leads to more tuples after updating.