Fix GPDB_12_MERGE_FIXME in nodeShareInputScan (#17138)
```
/* GPDB_12_MERGE_FIXME: would be nicer to store this hash in the DSM segment or DSA */
shareinput_Xslice_hash = ShmemInitHash("ShareInputScan notifications",
								  N_SHAREINPUT_SLOTS(),
								  N_SHAREINPUT_SLOTS(),
					      		          &info,
								  HASH_ELEM | HASH_BLOBS);
```
Shmem is static shared memory which is pre-allocated and initialized during the startup of the
postmaster. The address of static shared memory is the same between different backends.
It stores frequently accessed fixed-size data. Using static shared memory is much simpler.

DSM and DSA is dynamic shared memory which is allocated and managed dynamically
during the runtime of the postgres. The address between different backends is different. 
We usually use a static shared memory to store dsm_handle and other infos to help
backends to find its address. It is suitable for temporary data storage. 

shareinput_Xslice_state is small and we access the field frequently,
Instead of using the static memory to store shared_scan_id, dsm_handle and other infos,
store the shareinput_Xslice_state directly in static memory is more convenient.

So remove the fixme.