Check orphaned files functions exclude idle sessions when doing safety check
GDD and other services like GPCC may always keep the connections to
coordinator or segments. As discussed in PR#16903, we should skip
checking the idle session for orphaned files detection to avoid
that we can't get a detection result due to the ERROR `There is a
client session running on one or more segment.Aborting...`.