[pg_dump] Dump SCHEDULE clause of Dynamic Tables.
Dynamic Table's SCHEDULE clause is stored in pg_task jobs.
Add it when a Dynamic Table is dumped.

Since the SCHEDULE clause is optional, there would be no
error if we forget it when dump a Dynamic Table info.
And a default SCHEDULE is added with the value of Macro:
DYNAMIC_TABLE_DEFAULT_REFRESH_INTERVAL

Authored-by: Zhang Mingli avamingli@gmail.com